+++ src/index.js
@@ -27,13 +27,9 @@ export * from './platform.js'
 export * from './utils'
 
 // Vue Components
-import Calendar from './components/Calendar.vue'
-import Dropdown from './components/Dropdown.vue'
-import ObjectForm from './components/ObjectForm.vue'
-import ObjectFormField from './components/ObjectFormField.vue'
-import TransitionPage from './components/TransitionPage.vue'
+export * from './components/export.js'
 
 // Mixins
 import ModalMixin from './semantic/mixins/modal.js'
 
-export { Calendar, Dropdown, ObjectForm, ObjectFormField, TransitionPage, ModalMixin }
+export { ModalMixin }